Add files via upload
Day 10 💫 : KNN(k-nearest neighbors)
 It  is a non-parametric, supervised learning classifier, which uses proximity to make classifications or predictions about the grouping of an individual data point. It is one of the popular and simplest classification and regression classifiers used in machine learning today.

Way of performing KNN:-
Step 1 : Choose the number K neighbors.(i.e. K=5 ,it is among the number which is perfect for finding neighbor.)
Step 2 : Take the nearest neighbors of the new data point, according to the euclidean distance .euclidean distance = √[(x2 – x1)^2 + (y2 – y1)^2]
Step 4: Among these K-neighbors, count the number of datapoints in each category.
step 4: Assign new data point to the category where you counted the most neighbors.

When is KNN Algorithm used?
a. We have properly labeled data. For example, if we are predicting someone is having diabetes or not the final label can be 1 or 0. It cannot be NaN or -1.
b. Data is noise-free. For the diabetes data set we cannot have a Glucose level as 0 or 10000. It’s practically impossible.
c. Small dataset.

Work done in the given code snippet:-
performed 2 classification method i.e. Logistic regression and K-NN and finding which is more accurate with graph. In the given data and according to my result logistic regression performed better.
Source:- sklearn.org matplotlib.org(listedcolor) and some online websites
KNN is more suitable for non-linear relationships, small to medium datasets , realtime decision making but it is more expensive.